% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{get_random_name}
\alias{get_random_name}
\title{Generate random name.}
\usage{
get_random_name()
}
\description{
\code{get_random_name} generates a random name from the list of adjectives
and surnames in this package formatted as \code{"adjective_surname"}. For
example \code{'focused_turing'}.
}
